docker/kaggle_discord_bot_for_cloud_run/requirements.txt
uvicorn
fastapi
python-dotenv
kaggle
python-i18n[YAML]
pytz
